+++ eyes.appium.java/src/main/java/com/applitools/eyes/appium/Eyes.java
@@ -495,6 +495,7 @@ public class Eyes extends EyesBase {
 
         ValidationResult validationResult = new ValidationResult();
         validationResult.setAsExpected(result.getAsExpected());
+        ((AppiumScrollPositionProvider) getPositionProvider()).cleanupCachedData();
     }
 
     @Override